tutorial/GMM_toolBox/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
cffi==1.11.5
clickhouse-driver==0.0.19
cryptography==2.2.2
cx-Oracle==6.4.1
cycler==0.10.0
DBUtils==1.2
Django==2.0.4
django-celery==3.2.2
django-cors-headers==2.2.0
django-filter==2.0.0
django-filters==0.2.1
django-picklefield==1.1.0
django-vue==0.0.1
djangorestframework==3.8.2
eventlet==0.23.0
greenlet==0.4.13
idna==2.7
kiwisolver==1.0.1
kombu==3.0.37
Markdown==2.6.11
matplotlib==2.2.3
mysql-connector-python==8.0.11
MySQL-python==1.2.5
mysqlclient==1.3.12
npm==0.1.1
numpy==1.15.1
optional-django==0.1.0
ordereddict==1.1
pika==1.0.1
protobuf==3.6.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
PyMysqlDB==0.0.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
redis==2.10.6
roman==3.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.11.0
sklearn==0.0
vine==1.1.4
